refactor: expose components (#67)
