+++ src/index.ts
@@ -1,1 +1,11 @@
-export * as api from './api';
+import AccountManager from './AccountManager';
+import RelayPricer from './RelayPricer';
+import RelayClient from './RelayClient';
+
+export * from './api';
+export * from './discovery';
+export * from './gasEstimator';
+export * from './utils';
+export * from './common';
+
+export { AccountManager, RelayPricer, RelayClient }
